Move the zeroing function out to its own file.
