+++ stories/utils/adsense-data-zeroing.js
@@ -1,0 +1,26 @@
+export const zeroing = ( report ) => {
+	const zeroValue = ( cell ) => ( { ...cell, value: 0 } );
+
+	let clonedReport = { ...report };
+
+	const { totals, rows } = clonedReport;
+	const { cells } = totals;
+
+	clonedReport = {
+		...clonedReport,
+		totals: {
+			cells: cells.map( zeroValue ),
+		},
+		rows: rows.map( ( row ) => ( {
+			...row,
+			cells: row.cells.map( ( cell, index ) => {
+				if ( index !== 0 ) {
+					return zeroValue( cell );
+				}
+				return cell;
+			} ),
+		} ) ),
+	};
+
+	return clonedReport;
+};
